Refactored structure. Added owner field in PropertyItem. Deleted entity OwnershipHistoryService.java
