+++ src/main/java/com/example/collateralmanagement/models/entities/business/Department.java
@@ -1,6 +1,6 @@
-package com.example.collateralmanagement.domain.entities.business;
+package com.example.collateralmanagement.models.entities.business;
 
-import com.example.collateralmanagement.domain.entities.enums.DepartmentEnum;
+import com.example.collateralmanagement.models.entities.enums.DepartmentEnum;
 import jakarta.persistence.*;
 
 @Entity
