+++ src/main/java/com/example/collateralmanagement/repositories/AcquiredAssetRepository.java
@@ -1,6 +1,6 @@
 package com.example.collateralmanagement.repositories;
 
-import com.example.collateralmanagement.domain.entities.property.AcquiredAsset;
+import com.example.collateralmanagement.models.entities.property.AcquiredAsset;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
