+++ src/main/java/com/example/collateralmanagement/repositories/AppraisalCompanyRepository.java
@@ -1,6 +1,6 @@
 package com.example.collateralmanagement.repositories;
 
-import com.example.collateralmanagement.domain.entities.valuation.AppraisalCompany;
+import com.example.collateralmanagement.models.entities.valuation.AppraisalCompany;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
