+++ src/main/java/com/example/collateralmanagement/repositories/AppraisalRepository.java
@@ -1,6 +1,6 @@
 package com.example.collateralmanagement.repositories;
 
-import com.example.collateralmanagement.domain.entities.valuation.Appraisal;
+import com.example.collateralmanagement.models.entities.valuation.Appraisal;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
