+++ src/main/java/com/example/collateralmanagement/repositories/BankClientRepository.java
@@ -1,6 +1,6 @@
 package com.example.collateralmanagement.repositories;
 
-import com.example.collateralmanagement.domain.entities.business.BankClient;
+import com.example.collateralmanagement.models.entities.business.BankClient;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
