+++ src/main/java/com/example/collateralmanagement/repositories/CollateralRepository.java
@@ -1,6 +1,6 @@
 package com.example.collateralmanagement.repositories;
 
-import com.example.collateralmanagement.domain.entities.property.Collateral;
+import com.example.collateralmanagement.models.entities.property.Collateral;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
