+++ src/main/java/com/example/collateralmanagement/repositories/CollateralTypeRepository.java
@@ -1,6 +1,6 @@
 package com.example.collateralmanagement.repositories;
 
-import com.example.collateralmanagement.domain.entities.property.CollateralType;
+import com.example.collateralmanagement.models.entities.property.CollateralType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
