+++ src/main/java/com/example/collateralmanagement/repositories/LoanRepository.java
@@ -1,6 +1,6 @@
 package com.example.collateralmanagement.repositories;
 
-import com.example.collateralmanagement.domain.entities.business.Loan;
+import com.example.collateralmanagement.models.entities.business.Loan;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
