+++ src/main/java/com/example/collateralmanagement/repositories/MarketStandardsPerYearRepository.java
@@ -1,6 +1,6 @@
 package com.example.collateralmanagement.repositories;
 
-import com.example.collateralmanagement.domain.entities.valuation.MarketStandardsPerYear;
+import com.example.collateralmanagement.models.entities.valuation.MarketStandardsPerYear;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
