+++ src/main/java/com/example/collateralmanagement/repositories/MarketStandardsRevaluationRepository.java
@@ -1,6 +1,6 @@
 package com.example.collateralmanagement.repositories;
 
-import com.example.collateralmanagement.domain.entities.valuation.MarketStandardsRevaluation;
+import com.example.collateralmanagement.models.entities.valuation.MarketStandardsRevaluation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
