+++ src/main/java/com/example/collateralmanagement/repositories/PropertyItemRepository.java
@@ -1,6 +1,6 @@
 package com.example.collateralmanagement.repositories;
 
-import com.example.collateralmanagement.domain.entities.property.PropertyItem;
+import com.example.collateralmanagement.models.entities.property.PropertyItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
