+++ src/main/java/com/example/collateralmanagement/services/BankClientService.java
@@ -1,7 +1,7 @@
 package com.example.collateralmanagement.services;
 
 
-import com.example.collateralmanagement.domain.dtos.business.ImportClientDTO;
+import com.example.collateralmanagement.models.dtos.business.ImportClientDTO;
 
 public interface BankClientService {
