+++ src/main/java/com/example/collateralmanagement/services/LoanService.java
@@ -1,6 +1,6 @@
 package com.example.collateralmanagement.services;
 
-import com.example.collateralmanagement.domain.dtos.business.ImportLoanDTO;
+import com.example.collateralmanagement.models.dtos.business.ImportLoanDTO;
 
 public interface LoanService {
