--- src/main/java/com/example/collateralmanagement/services/OwnershipHistoryService.java
@@ -1,4 +1,0 @@
-package com.example.collateralmanagement.services;
-
-public interface OwnershipHistoryService {
-}
