+++ src/main/java/com/example/collateralmanagement/services/impl/businessServices/BankClientServiceImpl.java
@@ -1,7 +1,7 @@
 package com.example.collateralmanagement.services.impl.businessServices;
 
-import com.example.collateralmanagement.domain.dtos.business.ImportClientDTO;
-import com.example.collateralmanagement.domain.entities.business.BankClient;
+import com.example.collateralmanagement.models.dtos.business.ImportClientDTO;
+import com.example.collateralmanagement.models.entities.business.BankClient;
 import com.example.collateralmanagement.repositories.BankClientRepository;
 import com.example.collateralmanagement.services.BankClientService;
 import org.modelmapper.ModelMapper;
